[darwin] move ios windowing into seperate folder
