+++ xbmc/windowing/ios/CMakeLists.txt
@@ -1,0 +1,8 @@
+set(SOURCES WinEventsIOS.mm
+            WinSystemIOS.mm
+            VideoSyncIos.cpp)
+set(HEADERS WinEventsIOS.h
+            WinSystemIOS.h
+            VideoSyncIos.h)
+
+core_add_library(windowing_ios)
