+++ xbmc/windowing/ios/WinEventsIOS.h
