+++ xbmc/windowing/osx/CMakeLists.txt
@@ -1,25 +1,14 @@
-if(CORE_SYSTEM_NAME STREQUAL osx)
-  set(SOURCES OSScreenSaverOSX.cpp
-              WinEventsOSX.mm
-              WinEventsSDL.cpp
-              WinSystemOSX.mm
-              WinSystemOSXGL.mm
-              VideoSyncOsx.cpp)
-  set(HEADERS OSScreenSaverOSX.h
-              WinEventsOSX.h
-              WinEventsSDL.h
-              WinSystemOSX.h
-              WinSystemOSXGL.h
-              VideoSyncOsx.h)
-endif()
-
-if(CORE_SYSTEM_NAME STREQUAL ios)
-  set(SOURCES WinEventsIOS.mm
-              WinSystemIOS.mm
-              VideoSyncIos.cpp)
-  set(HEADERS WinEventsIOS.h
-              WinSystemIOS.h
-              VideoSyncIos.h)
-endif()
+set(SOURCES OSScreenSaverOSX.cpp
+            WinEventsOSX.mm
+            WinEventsSDL.cpp
+            WinSystemOSX.mm
+            WinSystemOSXGL.mm
+            VideoSyncOsx.cpp)
+set(HEADERS OSScreenSaverOSX.h
+            WinEventsOSX.h
+            WinEventsSDL.h
+            WinSystemOSX.h
+            WinSystemOSXGL.h
+            VideoSyncOsx.h)
 
 core_add_library(windowing_osx)
